third_party/WebKit/LayoutTests/fast/css/link-alternate-style-recalc-expected.txt

Check that alternate stylesheet load completion did not cause a synchronous style recalc.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS internals.updateStyleAndReturnAffectedElementCount() is not 0
PASS getComputedStyle(document.getElementById('t'), null).color is 'rgb(0, 0, 0)'
